+++ web/cypress/integration/route/create-route-with-proxy-rewrite-plugin.spec.js
@@ -22,9 +22,9 @@ import routeLocaleUS from '../../../src/pages/Route/locales/en-US';
 context('create route with proxy-rewrite plugin', () => {
   const selector = {
     name: '#name',
-    nodes_0_host: '#nodes_0_host',
-    nodes_0_port: '#nodes_0_port',
-    nodes_0_weight: '#nodes_0_weight',
+    nodes_0_host: '#submitNodes_0_host',
+    nodes_0_port: '#submitNodes_0_port',
+    nodes_0_weight: '#submitNodes_0_weight',
     nameSelector: '[title=Name]',
     deleteAlert: '.ant-modal-body',
     notification: '.ant-notification-notice-message',