+++ web/src/components/Plugin/UI/proxy-mirror.tsx
@@ -21,6 +21,7 @@ import { useIntl } from 'umi';
 
 type Props = {
   form: FormInstance;
+  schema: Record<string, any> | undefined;
 };
 
 const FORM_ITEM_LAYOUT = {
@@ -32,8 +33,9 @@ const FORM_ITEM_LAYOUT = {
   },
 };
 
-const ProxyMirror: React.FC<Props> = ({ form }) => {
+const ProxyMirror: React.FC<Props> = ({ form, schema }) => {
   const { formatMessage } = useIntl();
+  const properties = schema?.properties;
 
   return (
     <Form form={form} {...FORM_ITEM_LAYOUT}>
@@ -44,10 +46,7 @@ const ProxyMirror: React.FC<Props> = ({ form }) => {
         tooltip={formatMessage({ id: 'component.pluginForm.proxy-mirror.host.tooltip' })}
         rules={[
           {
-            pattern: new RegExp(
-              /^http(s)?:\/\/[a-zA-Z0-9][-a-zA-Z0-9]{0,62}(\.[a-zA-Z0-9][-a-zA-Z0-9]{0,62})+(:[0-9]{1,5})?$/,
-              'g',
-            ),
+            pattern: new RegExp(`${properties.host.pattern}`, 'g'),
             message: formatMessage({ id: 'component.pluginForm.proxy-mirror.host.ruletip' }),
           },
         ]}